Use size_t for graph indicies

+++ test/test_list_graph.cpp
@@ -6,43 +6,43 @@ namespace {
 
 class ListGraphUndirectedTest : public ::testing::Test {
 protected:
-    simple_graph::ListGraph<false, int, int> g;
+    simple_graph::ListGraph<false, size_t, size_t> g;
 };
 
 TEST_F(ListGraphUndirectedTest, test_add_vertex)
 {
     EXPECT_EQ(0, g.vertex_num());
-    g.add_vertex(simple_graph::Vertex<int>(2));
+    g.add_vertex(simple_graph::Vertex<size_t>(2));
     EXPECT_EQ(3, g.vertex_num());
-    EXPECT_EQ(-1, g.add_vertex(simple_graph::Vertex<int>(2)));
+    EXPECT_EQ(-1, g.add_vertex(simple_graph::Vertex<size_t>(2)));
 }
 
 TEST_F(ListGraphUndirectedTest, test_set_vertex)
 {
     EXPECT_EQ(0, g.vertex_num());
 
-    g.set_vertex(simple_graph::Vertex<int>(1, 4));
+    g.set_vertex(simple_graph::Vertex<size_t>(1, 4));
     EXPECT_EQ(2, g.vertex_num());
     EXPECT_EQ(4, g.vertex(1).data());
 
-    g.set_vertex(simple_graph::Vertex<int>(1, 5));
+    g.set_vertex(simple_graph::Vertex<size_t>(1, 5));
     EXPECT_EQ(2, g.vertex_num());
     EXPECT_EQ(5, g.vertex(1).data());
 }
 
 TEST_F(ListGraphUndirectedTest, test_add_edge)
 {
-    for (int i = 0; i < 4; ++i) {
-        g.set_vertex(simple_graph::Vertex<int>(i, i));
+    for (size_t i = 0; i < 4; ++i) {
+        g.set_vertex(simple_graph::Vertex<size_t>(i, i));
     }
 
     EXPECT_EQ(4, g.vertex_num());
 
-    g.add_edge(simple_graph::Edge<int>(0, 1, 11));
-    g.add_edge(simple_graph::Edge<int>(1, 2, 12));
-    g.add_edge(simple_graph::Edge<int>(2, 3, 13));
+    g.add_edge(simple_graph::Edge<size_t>(0, 1, 11));
+    g.add_edge(simple_graph::Edge<size_t>(1, 2, 12));
+    g.add_edge(simple_graph::Edge<size_t>(2, 3, 13));
 
-    simple_graph::Edge<int> e = g.edge(1, 2);
+    simple_graph::Edge<size_t> e = g.edge(1, 2);
     EXPECT_EQ(12, e.weight());
 
     int expected_edges[3][3] = {{0, 1, 11}, {1, 2, 12}, {2, 3, 13}};
@@ -57,37 +57,37 @@ TEST_F(ListGraphUndirectedTest, test_add_edge)
 
 TEST_F(ListGraphUndirectedTest, test_bfs_direct_order)
 {
-    for (int i = 0; i < 4; ++i) {
-        g.set_vertex(simple_graph::Vertex<int>(i, i));
+    for (size_t i = 0; i < 4; ++i) {
+        g.set_vertex(simple_graph::Vertex<size_t>(i, i));
     }
 
     EXPECT_EQ(4, g.vertex_num());
 
-    g.add_edge(simple_graph::Edge<int>(0, 1));
-    g.add_edge(simple_graph::Edge<int>(1, 2));
-    g.add_edge(simple_graph::Edge<int>(2, 3));
+    g.add_edge(simple_graph::Edge<size_t>(0, 1));
+    g.add_edge(simple_graph::Edge<size_t>(1, 2));
+    g.add_edge(simple_graph::Edge<size_t>(2, 3));
 
-    std::vector<int> path;
-    int p = 3;
-    std::function<bool(int)> f = [&](int c) { return c == p; };
+    std::vector<size_t> path;
+    size_t p = 3;
+    std::function<bool(size_t)> f = [&](size_t c) { return c == p; };
     EXPECT_EQ(true, bfs(g, 0, f, &path));
 }
 
 TEST_F(ListGraphUndirectedTest, test_bfs_reverse_order)
 {
     for (int i = 0; i < 4; ++i) {
-        g.set_vertex(simple_graph::Vertex<int>(i, i));
+        g.set_vertex(simple_graph::Vertex<size_t>(i, i));
     }
 
     EXPECT_EQ(4, g.vertex_num());
 
-    g.add_edge(simple_graph::Edge<int>(0, 1));
-    g.add_edge(simple_graph::Edge<int>(1, 2));
-    g.add_edge(simple_graph::Edge<int>(2, 3));
+    g.add_edge(simple_graph::Edge<size_t>(0, 1));
+    g.add_edge(simple_graph::Edge<size_t>(1, 2));
+    g.add_edge(simple_graph::Edge<size_t>(2, 3));
 
-    std::vector<int> path;
-    int p = 0;
-    std::function<bool(int)> f = [&](int c) { return c == p; };
+    std::vector<size_t> path;
+    size_t p = 0;
+    std::function<bool(size_t)> f = [&](size_t c) { return c == p; };
     EXPECT_EQ(true, bfs(g, 3, f, &path));
 }
